requirements.txt
fastapi==0.75.0
matplotlib==3.5.1
numpy==1.22.3
Pillow==9.0.1
requests==2.26.0
scikit_image==0.19.2
scikit_learn==1.0.2
seaborn==0.11.2
skimage==0.0
